// tmp/src/app/model/manager-device-group-interface.ts
export interface ManagerDeviceGroup {
  groupId?: string;
  thingId?: string;
  managerName?: string;
  creatorManagerEmail?: string;
  groupManagementId?: string;
  groupManagerEmail?: string;
  groupName?: string;
  createdAt?: string;
  deviceCount?: number;
  memberCount?: number;
  canEdit: boolean;
}